Remove RequestHandler in favor of Router
Deleted:
- uncors_handler_test.go - Tests for RequestHandler (replaced by router_test.go)
- handler_routes_test.go - Tests using RequestHandler
- har_middleware.go - Methods specific to RequestHandler

Modified:
- router.go: Added setDefaultHandler() and errHostNotMapped moved from uncors_handler
- routes.go: Removed RequestHandler methods, kept generic registerMatchedRoutes helper
- uncors_handler.go: Removed RequestHandler class and related options, kept type definitions
- router_test.go: Updated with factory functions and test data from uncors_handler_test

Router is now the primary handler implementation with complete test coverage
matching the original RequestHandler functionality.
